+++ packages/core/helper-plugin/MIGRATION_GUIDE.md
@@ -240,6 +240,10 @@ const MyComponent = (props) => {
 
 This component has been removed and not replaced. If you feel like you need this component, please open an issue on the Strapi repository to discuss your usecase.
 
+### RelativeTime
+
+This component has been removed and not replaced. If you feel like you need this component, please open an issue on the Strapi repository to discuss your usecase.
+
 ### SettingsPageTitle
 
 This component has been removed and not replaced. If you feel like you need this component, please open an issue on the Strapi repository to discuss your usecase.
